test(core): move utxo test utils to 'util/'
Issue: BG-38773
